Split scheduling infra into a dedicated mod.
